+++ lib/append.js
@@ -1,8 +1,6 @@
-import rdf from 'rdf-ext'
 import { Transform } from 'readable-stream'
-import { localFetch } from './localFetch/localFetch.js'
-import { applyOptions } from './metadata/applyOptions.js'
-import getDatasetBuilder from './metadata/datasetBuilder'
+import { applyReplacements } from './metadata/applyReplacements.js'
+import getDatasetBuilder from './metadata/datasetTemplateBuilder.js'
 
 class PushDatasetOnFlush extends Transform {
   constructor (options) {
@@ -42,7 +40,8 @@ async function createAppend ({
   const pointer = rootTerm ? rootTerm.term : input.term
   const base = this?.basePath ? this.basePath : basePath
 
-  const datasetBuilder = getDatasetBuilder({ input, base, pointer }, { dateModified, dateCreated, graph })
+  const replacer = (dataset, metadata) => applyReplacements(dataset, metadata, { dateModified, dateCreated, graph })
+  const datasetBuilder = await getDatasetBuilder({ input, base, pointer }, replacer)
 
   return new PushDatasetOnFlush({
     datasetBuilder
