+++ lib/appendMany.js
@@ -1,8 +1,8 @@
 import rdf from 'rdf-ext'
 import { Transform } from 'readable-stream'
-import getDatasetBuilder from './metadata/datasetBuilder.js'
-
+import { applyReplacements } from './metadata/applyReplacements'
 import { wellKnownDatasetClasses } from './metadata/datasetClasses.js'
+import getDatasetBuilder from './metadata/datasetTemplateBuilder.js'
 
 const rdfType = rdf.namedNode('http://www.w3.org/1999/02/22-rdf-syntax-ns#type')
 
@@ -67,7 +67,8 @@ async function createAppendMany ({
   const pointer = rootTerm ? rootTerm.term : input.term
   const base = this?.basePath ? this.basePath : basePath
 
-  const datasetBuilder = getDatasetBuilder({ input, base, pointer }, { dateModified, dateCreated, graph })
+  const replacer = (dataset, metadata) => applyReplacements(dataset, metadata, { dateModified, dateCreated, graph })
+  const datasetBuilder = await getDatasetBuilder({ input, base, pointer }, replacer)
 
   return new PushForSubjectsOfTypes({
     types,
