+++ lib/metadata/applyReplacements.js
@@ -43,7 +43,7 @@ function resolveNamedDate (value, metadata) {
   return namedDateLiterals.has(value) ? namedDateLiterals.get(value)(metadata) : toDateLiteral(value)
 }
 
-function applyOptions (dataset, metadata = {}, options = {}) {
+function applyReplacements (dataset, metadata = {}, options = {}) {
   // dateModified
   if (options.dateModified) {
     const dateModifiedLiteral = resolveNamedDate(options.dateModified, metadata)
@@ -66,4 +66,4 @@ function applyOptions (dataset, metadata = {}, options = {}) {
   return dataset
 }
 
-export { applyOptions }
+export { applyReplacements }
