+++ lib/metadata/datasetTemplateBuilder.js
@@ -1,13 +1,13 @@
+import TraverserFactory from '@rdfjs/traverser/Factory.js'
 import rdf from 'rdf-ext'
-import { localFetch } from '../localFetch/localFetch'
+import { localFetch } from '../localFetch/localFetch.js'
 import * as ns from '../namespaces.js'
-import { applyOptions } from './applyOptions'
 
 function isClownface (input) {
   return input.term
 }
 
-const boundedDescription = rdf.traverser(({
+const boundedDescription = new TraverserFactory().traverser(({
   dataset,
   level,
   quad
@@ -21,11 +21,11 @@ async function readFromLocal (input, basePath, pointer) {
   const { quadStream, metadata } = await localFetch(input, basePath)
   const dataset = await rdf.dataset().import(quadStream)
 
-  return pointer ? {
-      dataset: boundedDescription.match(rdf.clownface({ dataset: dataset, term: rootTerm.term })), metadata
-    }
+  return pointer
+    ? {
+        dataset: boundedDescription.match(rdf.clownface({ dataset: dataset, term: pointer })), metadata
+      }
     : { dataset: dataset, metadata }
-
 }
 
 function getBuilder (dataset, pointer) {
@@ -41,15 +41,12 @@ function getBuilder (dataset, pointer) {
   return datasetBuilder
 }
 
-async function getDatasetBuilder ({ input, base, pointer }, replacements) {
+async function getDatasetTemplateBuilder ({ input, base, pointer }, replacer = (dataset, metadata) => dataset) {
   const {
     dataset,
     metadata
   } = isClownface(input) ? await readFromClownface(input) : await readFromLocal(input, base, pointer)
-
-
-  const datasetWithReplacements = applyOptions(dataset, metadata, replacements)
-  return getBuilder(datasetWithReplacements, pointer)
+  return getBuilder(replacer(dataset, metadata), pointer)
 }
 
-export default getDatasetBuilder
+export default getDatasetTemplateBuilder
