+++ package.json
@@ -25,6 +25,7 @@
     "@rdfjs/namespace": "^1.1.0",
     "@rdfjs/term-map": "^1.0.0",
     "@rdfjs/term-set": "^1.0.1",
+    "@rdfjs/traverser": "^0.1.0",
     "clownface": "^1.3.0",
     "file-fetch": "^1.7.0",
     "lodash": "^4.17.21",
